fixed sprintf param format in Log::logR

+++ Log.cpp
@@ -243,14 +243,11 @@ bool Log::logR ( int64_t now, int32_t type, const char *msg, bool forced ) {
         else
         {
             if ( g_hostdb.getNumHosts() <= 999 )
-                    p += sprintf ( p , "%" PRIu64" %03" PRId32" ",
-                              now , g_hostdb.m_hostId );
+                    p += sprintf ( p , "%" PRIu64 " %03" PRId32 " ", (uint64_t)now , g_hostdb.m_hostId );
             else if ( g_hostdb.getNumHosts() <= 9999 )
-                    p += sprintf ( p , "%" PRIu64" %04" PRId32" ",
-                              now , g_hostdb.m_hostId );
+                    p += sprintf ( p , "%" PRIu64" %04" PRId32" ", (uint64_t)now , g_hostdb.m_hostId );
             else if ( g_hostdb.getNumHosts() <= 99999 )
-                    p += sprintf ( p , "%" PRIu64" %05" PRId32" ",
-                              now , g_hostdb.m_hostId );
+                    p += sprintf ( p , "%" PRIu64" %05" PRId32" ", (uint64_t)now , g_hostdb.m_hostId );
         }
 	}
